Added some 64bit warning fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@23612 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/archdep.c
@@ -142,7 +142,7 @@ char *archdep_program_name(void)
             e = argv0 + strlen(argv0);
         }
 
-        len = e - s + 1;
+        len = (int)(e - s + 1);
         program_name = lib_malloc(len);
         memcpy(program_name, s, len - 1);
         program_name[len - 1] = 0;
