+++ vice/src/cbmdos.c
@@ -227,7 +227,7 @@ unsigned int cbmdos_command_parse(cbmdos_cmd_parse_t *cmd_parse)
         p++;
 #endif
 
-    cmdlen = cmd_parse->cmdlength - (p - cmd_parse->cmd);
+    cmdlen = cmd_parse->cmdlength - (int)(p - cmd_parse->cmd);
     cmd_parse->parselength = 0;
 
     /* Temporary hack.  */
@@ -303,7 +303,7 @@ unsigned int cbmdos_command_parse(cbmdos_cmd_parse_t *cmd_parse)
 
         c = (char *)memchr(p, ',', cmdlen);
         if (c) {
-            cmdlen -= (c - (const char *)p);
+            cmdlen -= (int)(c - (const char *)p);
             p = (BYTE *)c;
         } else {
             cmdlen = 0;
